3.1.1 Video Introduction - Standard API's
From what we’ve learned so far, it is time now to enjoy some of the standard API’s your browser provides. 
This module is centered around the BOM; the “Browser Object Model”. That provides standard objects such as the navigator or the window object. We already had a glimpse of these, this time we look at them more into details. In addition, we will look at some HTML5 standards APIs provided by your browser.