3.2.3 Iterating on array elements
Hello! Let's look at the different methods you can use to iterate on each element of an array. One of the most practical one consists in using the “forEach” iterator we already met during some previous examples, in particular during Module 1. 
The syntax is this one, you start with the name of the array, and you type “forEach” with a capital “E”. And this method takes a single callback. Usually, we type the callback body between the parenthesis. Declaring an anonymous function, an anonymous callback. And the first example is to use only a single argument that will be the current element while we are iterating. I call it “day” here, and day will be: Monday, Tuesday, Wednesday,… each time, we iterate on the collection of the days, on the array of days.