3.5.4 [Advanced] a multiple image, sound and music loader
Live coding transcript: a multiple image, sound and music loader
I prepared for you... more for advanced users... a multiple image, sound and music loader.
It's a function that you will call... that will load in background, for you, images, sound samples and also deal with streamed music. And when everything is ready, it will call a function of your own that will start the game.