4.1.1 Introduction JS objects
This module is about JavaScript objects.
So far, we’ve only scratched the surface! And it’s time to get into detail. As with many other programming languages, JavaScript objects can be compared to real-life objects. A JavaScript object is an entity with a type and properties that define its characteristics. It also comes with special functions associated to it called methods.