2.2.2 Conditional statements
Live coding transcript: switch statement.
I prepared a small example in which we take a day number between 0 and 6, 0 for Monday, 1 for Tuesday, and so on…And to pick a random number that is an integer, we used a formula like this, that uses a predefined object called Math.
Math.random() returns a number between 0 and 1, 6 multiplied by this, will give a number between 0 and 7, and we just return the rounded nearest integer. When we execute this, "picked day number..." you see that it takes some random values each time we execute.