2.4.4  Page lifecycle events
Live coding transcript: page ‘load’ event and the event object.
One of the events you will use is called the 'load' event, and you can apply it on the page using onload equal followed by the name of the callback.
I usually call this callback "init" and this is a good practice: always use a function that will be called when the page is loaded. I'm adding a console.log message just to verify.