4.2.6 "this": accessing properties
Live coding transcript: add methods to the player object from the game.
Let's have a look now at the game we developed during the first three modules. That's with a small red squared player that follow the mouse. We had an animation loop that 60 times/second cleared the canvas, draws the player (the red square), moved the player, following the mouse, and then asked for a new frame of animation.