5.2.1 References and objects
References and objects:
When you define a variable, this is what happens: if its value is a primitive value, a number, a string or a boolean for example, the variable contains this value directly. On the opposite, if its value is an object with braces or created with the “new” keyword we saw during module 4, the variable contains the memory address of the object. We say that "this variable points to an object" or "references this object". And accessing the variable will automatically resolve the reference, meaning that the value of the variable is the referenced object.