5.2.4 Built-in JS class: Object
JavaScript comes with a set of predefined objects or built-in objects such as Math, String, Date and so on. For developers that already know the Java language... or C#, you can imagine this more like a class from the developer's kit, a class that is provided by the language itself. The consequence of this is that on all the objects you manipulate in JavaScript, you will be able to use a predefined set of methods and properties that will be inherited from this predefined object.